% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cces18_samp}
\alias{cces18_samp}
\title{Sample CCES}
\format{
\describe{
\item{case_id}{Unique CCES respondent identifier}
\item{st}{State abbreviation}
\item{cd}{congressional district (for US House)}
\item{pid}{Partisan identification. Leaners based on pid7 in the CCES
are coded as partisans}
\item{race}{Race as coded by the cumulative common content}
\item{newsint}{News interest  as coded by the cumulative common content}
\item{ideo5}{Ideology as coded by the cumulative common content}
\item{PRS_party, USS_party, GOV_party, USH_party, ATG_party,  SOS_party}{
Party vote for several offices. PRS = President, USS = US Senate,
GOV = Governor, USH = US House, ATG = Attorney General, SOS = Secretary of State}
}
}
\usage{
cces18_samp
}
\description{
Sample from the CCES 2018, Massachusetts, Wisconsin, and Michigan.
Includes vote choice for President, Governor, Congress, Attorney General, Secreatary
of State, and demographics, which were all up for election in these states.
President is retrospective from 2016.
}
\examples{
cces18_samp
}
\keyword{datasets}
