% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterCVR.R
\name{cl_Mstep_mu_bsl}
\alias{cl_Mstep_mu_bsl}
\title{compute mu for cluster k and office j}
\usage{
cl_Mstep_mu_bsl(y_j, zeta_k, n_u, l_mat)
}
\arguments{
\item{y_j}{vector of complete y at office j}

\item{zeta_k}{responsibility parameters, for each observation, for cluster k,}

\item{n_u}{vector of sizes}

\item{l_mat}{A constant N by (L+1) matrix to transform the y_j vector into
a matrix of indicators}
}
\value{
A L + 1 -length simplex mu
}
\description{
compute mu for cluster k and office j
}
\keyword{internal}
