% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterCVR.R
\name{cl_Mstep_mu_IIA}
\alias{cl_Mstep_mu_IIA}
\title{compute mu with a custom multinomial logit}
\usage{
cl_Mstep_mu_IIA(psi_init, y_j, zeta_k, m_j_fmt)
}
\arguments{
\item{psi_init}{L psi parameters (for coefficients excluding baseline abstain)}

\item{y_j}{vector of complete y at office j}

\item{zeta_k}{responsibility parameters, for each observation, for cluster k,}

\item{m_j_fmt}{missingness vector organized in particular way for optim}
}
\description{
compute mu with a custom multinomial logit
}
\keyword{internal}
