% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterCVR.R
\name{cl_Mstep_pi}
\alias{cl_Mstep_pi}
\title{Estep (MLE) for pi}
\usage{
cl_Mstep_pi(zeta, n_u, X = NULL, covs_gamma = NULL)
}
\arguments{
\item{zeta}{A N by K matrix of responsibility parameters}

\item{n_u}{A N by 1 vector for the number of observations for each profile.
u indexes "unique" profiles, instead of observations. If \code{fast} = \code{FALSE},
and the observations are simply voters, the value is 1 for everyone}

\item{X}{covariates, if available.}

\item{covs_gamma}{starting coefficients from \code{emlogit(zeta, X)}.}
}
\value{
A K by 1 vector of the mixing proportion pi (pi)
}
\description{
Estep (MLE) for pi
}
\keyword{internal}
