% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_clusters.R
\name{fmt_mu_viz}
\alias{fmt_mu_viz}
\title{Setup and visualize cluster characteristics}
\usage{
fmt_mu_viz(tbl)
}
\arguments{
\item{tbl}{output of get_mus}
}
\description{
\link{fmt_mu_viz} will do a few more formatting steps to prepare for visualization.
To actually run the graph, use the sample code in the examples below. I do not
make the sample code into a function for now because it likely differs case by case.
}
\examples{

\dontrun{
library(ggplot2)
data(simdata_miss)
cl_out  <- clusterCVR(simdata_miss, IIA = TRUE, init = "kmeans")

mu_out <- get_mus(cl_out)
mu_fmt <- fmt_mu_viz(mu_out)

ggplot(mu_fmt, aes(x =  mu, y = fct_rev(office), fill = fct_rev(vote))) +
  facet_wrap(~ cl_lbl) +
  geom_col(width = 0.5) +
  theme_classic() +
  guides(fill = guide_legend(name = "", reverse = TRUE)) +
  geom_text(aes(label = pct_lbl),
            position = position_stack(vjust = 0.5),
            size = 2.5,
            alpha = 0.8,
            color = "black",
            family = "mono") +
  theme(legend.position = "bottom",
        axis.text = element_text(color = "black"),
        axis.line.y = element_blank()) +
  labs(y = "",
       x = "Probability of Voting ...",
       fill = "") +
  scale_x_continuous(expand = c(0, 0)) +
  scale_fill_brewer(palette = "Greys") +
  theme(panel.spacing = unit(2, "lines"),
        legend.margin = unit(c(0, 0, 0, 0), "cm"),
        legend.box.spacing = unit(0.2, "lines"),
        legend.spacing = unit(0.2, "lines"))
}
}
