% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_clusters.R
\name{get_mus}
\alias{get_mus}
\alias{get_mu_ell}
\title{Extract the latent choice parameter in tidy format}
\usage{
get_mus(object)

get_mu_ell(outcome, data, names = office_names)
}
\arguments{
\item{object}{An object estimated by clusterCVR}
}
\value{
A tibble in tidy (long) format, where each row defines
a value of mu for a particular cluster, office, and vote choice. In the
current notation, there are \code{K} by \code{D} by \code{L + 1} observations.
}
\description{
Extract the latent choice parameter in tidy format

Get mu for ell
}
\details{
Currently it hard-codes the levels 1 as abstain, 2 as split, and 3
as straight.
}
\examples{
data(simdata_miss)
cl_out  <- clusterCVR(simdata_miss, IIA = TRUE, init = "kmeans")

mu_out <- get_mus(cl_out)


}
\keyword{internal}
