% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_mlogit.R
\name{grad_vclogit}
\alias{grad_vclogit}
\title{Evaluate gradient of the log likelihood}
\usage{
grad_vclogit(par, y_j, zeta_k, m_j)
}
\arguments{
\item{par}{The parameter of interest. A (L)-length vector of mlogit
coefficients for cluster k and office j. The coefficient for the reference
category, corresponding to abstention, is abbreviated here. That coefficient
is constrained to be 0. Therefore, the log likelihood appends 0 to the beginning
each time.}

\item{y_j}{N length vector of outcomes}

\item{zeta_k}{N-length vector of posterior for membership in cluster k}

\item{m_j}{N by (L + 1) matrix of missingness indicators}
}
\description{
Evaluate gradient of the log likelihood
}
