% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_diagnosis.R
\name{graph_trend}
\alias{graph_trend}
\title{Graph parameter fit and log likelihood}
\usage{
graph_trend(params_stacked, dat = data)
}
\arguments{
\item{params_stacked}{Output from \link{summ_params}}

\item{dat}{The input data for the original clusterCVR object}
}
\description{
EM algorithms should exhibit a monotonically increasing
log likelihood. This function plots the log likelihood
across iterations, as well as the change maximum change
in the parameter values.
}
