% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{loglik_diff}
\alias{loglik_diff}
\title{Check change in log likelihood}
\usage{
loglik_diff(loglik, loglik_pre)
}
\arguments{
\item{loglik}{ll at iteration it}

\item{loglik_pre}{ll at previous iteration}
}
\description{
Print a warning if not improving
}
\keyword{internal}
