% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{loglik_obs}
\alias{loglik_obs}
\title{Observed log likelihood}
\usage{
loglik_obs(iter, pi_it, mu_it, y, m, n_u, .IIA)
}
\arguments{
\item{iter}{the EM object with just one iteration}

\item{pi_it}{estimate of pi at iteration it}

\item{mu_it}{estimate of the mu array at iteration it}

\item{y}{A U by K matrix of the data}

\item{m}{A matrix of the missingness}

\item{n_u}{A vector of observation counts for each row in y}

\item{.IIA}{whether the model is using a multinomial logit.}
}
\value{
the total log likelihood
}
\description{
Observed log likelihood
}
