% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{print_iter}
\alias{print_iter}
\title{print iterations to console}
\usage{
print_iter(iter)
}
\arguments{
\item{iter}{Number of iterations}
}
\description{
print iterations to console
}
\keyword{internal}
