% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simdata_miss}
\alias{simdata_miss}
\title{Simulated data (some races not contested)}
\format{
A list with several parameters
\describe{
\item{y}{Numeric full matrix}
\item{uy}{Same numeric full matrix, unique profiles}
\item{m}{Indicator for the missingness in each cell of \code{y}. See writeup}
...
}
}
\usage{
simdata_miss
}
\description{
Simulated dataset that can be used for clustering
}
\keyword{datasets}
