% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_diagnosis.R
\name{summ_params}
\alias{summ_params}
\title{Stack all pre-post comparisons}
\usage{
summ_params(obj)
}
\arguments{
\item{obj}{An object of class \code{clusterCVR}}
}
\description{
Stack all pre-post comparisons
}
\examples{
data("simdata_full")
em_full_eq <- clusterCVR(simdata_full, init = "equal")

params_eq <- summ_params(em_full_eq)
graph_trend(params_eq, simdata_full)


}
\seealso{
\code{\link[=summary.clusterCVR]{summary.clusterCVR()}}, \code{\link[=graph_trend]{graph_trend()}}
}
