% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.clusterCVR}
\alias{summary.clusterCVR}
\title{Print summary of clusterCVR object}
\usage{
\method{summary}{clusterCVR}(x)
}
\arguments{
\item{x}{An object of class \link{clusterCVR}}
}
\description{
Print summary of clusterCVR object
}
